+++ src/meshapi/admin/__init__.py
@@ -1,4 +1,8 @@
 from .admin import *
+from .member import *
 from .building import *
 from .node import *
+from .device import *
+from .sector import *
+from .link import *
 from .inlines import *
